+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. -->
 - feat: Add `node-import/prefer-node-protocol` rule, requiring `node:` prefix
+- feat: Drop `@typescript-eslint/naming-convention`
 
 ## 8.5.0 - 8.5.1